fix: move test sorting function to correct dir
