+++ e2e-tests/Landing.spec.ts
@@ -1,10 +1,10 @@
 import { expect, test } from "@playwright/test";
 
 import { Offer } from "../src/lib/types/offer";
-import { sortOffersBy } from "../src/lib/utils/sort";
 import { assertOffer } from "./assert/offer";
 import { defaultMockOffers } from "./mocks/defaultMockOffers";
 import { mockSubgraph } from "./mocks/mockGetBase";
+import { sortOffersBy } from "./utils/sort";
 
 test.describe("Root page (Landing page)", () => {
   test.describe("Header & footer", () => {
