+++ e2e-tests/utils/sort.ts
@@ -1,4 +1,4 @@
-import { Offer } from "lib/types/offer";
+import { Offer } from "src/lib/types/offer";
 
 export const sortOffersBy =
   (by: { property: "name"; asc: boolean }) =>
